Move test LogicalPlanR[Fix] into frontend package.
