+++ connector/src/test/scala/quasar/qscript/Rewrite.scala
@@ -31,7 +31,7 @@ import pathy.Path._
 import scalaz._, Scalaz._
 
 class QScriptRewriteSpec extends quasar.Qspec with CompilerHelpers with QScriptHelpers {
-  import quasar.sql.fixpoint.lpf
+  import quasar.frontend.fixpoint.lpf
 
   val rewrite = new Rewrite[Fix]
