+++ core/src/test/scala/quasar/sql/compiler.scala
@@ -27,7 +27,7 @@ class CompilerSpec extends quasar.Qspec with CompilerHelpers {
   //     need to do a better job of handling this.
   import quasar.LogicalPlan._
   import quasar.std.StdLib._, relations._, StdLib.set._, string._, structural._
-  import quasar.sql.fixpoint.lpf
+  import quasar.frontend.fixpoint.lpf
 
   "compiler" should {
     "compile simple constant example 1" in {
